+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 3.1.0 2020-10-27
+### Added
+- `resolveReferences` method to `AbstractServiceFactoryExtension`.
+- `PassThroughCompiler` for simple services that do not require any special compiler logic.
+
 ## 3.0.4 2020-09-06
 ### Fixed
 - Allow numbers in class declarations.
@@ -69,7 +74,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Created the initial services layer.
 
-- [3.0.3 > Unreleased](https://github.com/ulrack/services/compare/3.0.3...HEAD)
+- [3.1.0 > Unreleased](https://github.com/ulrack/services/compare/3.1.0...HEAD)
+- [3.0.4 > 3.1.0](https://github.com/ulrack/services/compare/3.0.4...3.1.0)
+- [3.0.3 > 3.0.4](https://github.com/ulrack/services/compare/3.0.3...3.0.4)
 - [3.0.2 > 3.0.3](https://github.com/ulrack/services/compare/3.0.2...3.0.3)
 - [3.0.1 > 3.0.2](https://github.com/ulrack/services/compare/3.0.1...3.0.2)
 - [3.0.0 > 3.0.1](https://github.com/ulrack/services/compare/3.0.0...3.0.1)